backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.1.0
# OCR 相关依赖较重且对运行环境要求高；如仅想启动服务/看界面，可先不装并保持 ENABLE_OCR=false
# paddlepaddle==2.6.2
# paddleocr==2.7.3
opencv-python==4.6.0.66
# sentence-transformers 会拉取 torch 等大依赖；如只需启动服务，可先不装（向量搜索会降级）
# sentence-transformers>=2.3.0
faiss-cpu==1.12.0
numpy==1.26.4
google-generativeai==0.3.2
openai==1.12.0
python-dotenv==1.0.0
networkx==3.2.1
pyvis==0.3.2

